+++ components/_const.ts
@@ -1,6 +1,6 @@
 export const site = process.env.NEXT_PUBLIC_SITE_NAME;
 export const rule = process.env.NEXT_PUBLIC_RULE;
 export const shortRule = process.env.NEXT_PUBLIC_SHORT_RULE ?? "";
-export const recreateRule = process.env.NEXT_PUBLIC_RECREATE_RULE;
+export const recreateRule = process.env.NEXT_PUBLIC_RECREATE_RULE ?? "";
 
 export const answerLimit = process.env.NEXT_PUBLIC_ANSWER_LIMIT;
